requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastapi-lifespan-manager==0.1.4
fastapi-socketio==0.0.10
fastapi-swagger==0.2.5
fastapi-utils==0.8.0
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
generativeai==0.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-speech==2.28.1
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.15.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.7.0
langchain==0.3.14
langchain-community-0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langsmith==0.2.11
libclang==18.1.1
llama-llm==0.0.15
loguru==0.7.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
PyMuPDF==1.25.2
pyparsing==3.2.0
PyPDF2==3.0.1
python-configuration==0.12.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.10.1
python-jose==3.3.0
python-json-logger==3.2.1
python-socketio==5.11.4
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
